% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_csfiles_create.R
\name{kg_csfiles_create}
\alias{kg_csfiles_create}
\title{Create C# files from "KG_Shortcut.xlsx"}
\usage{
kg_csfiles_create(xlsxFile, lpname, savepath)
}
\arguments{
\item{xlsxFile}{An xlsx file.}

\item{lpname}{Loupedeck plugin name.}

\item{savepath}{Specify where to save CS files.}
}
\value{
C# Source Files
}
\description{
\code{kg_csfiles_create} This function create C# files from "KG_Shortcut.xlsx".
}
\examples{
kg_csfiles_create(xlsxFile = "../ExsampleData/KG_Shortcut.xlsx", savepath = getwd())
}
