% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_xlsx_color.R
\name{kg_xlsx_color}
\alias{kg_xlsx_color}
\title{Get icon color data from keyboard shortcut excel file}
\usage{
kg_xlsx_color(xlsxFile, fillcol = 2, textcol = 3, ncol = 6)
}
\arguments{
\item{xlsxFile}{An xlsx file.}

\item{fillcol}{fill color column Number.}

\item{textcol}{text color column Number.}

\item{ncol}{number of columns in the Excel file.}
}
\value{
Return fill and text color in a data frame
}
\description{
\code{kg_xlsx_color} This function Get icon color data from
    keyboard shortcut excel file.
}
\examples{
kg_xlsx_color(xlsxFile, fillcol = 2, textcol = 3, ncol = 6)
}
